+++ src/main/java/com/buuz135/simpleclaims/commands/subcommand/party/op/OpAddChunkAmountCommand.java
@@ -4,7 +4,6 @@ import com.buuz135.simpleclaims.claim.ClaimManager;
 import com.buuz135.simpleclaims.claim.party.PartyOverride;
 import com.buuz135.simpleclaims.claim.party.PartyOverrides;
 import com.buuz135.simpleclaims.commands.CommandMessages;
-import com.hypixel.hytale.protocol.GameMode;
 import com.hypixel.hytale.server.core.command.system.CommandContext;
 import com.hypixel.hytale.server.core.command.system.CommandSender;
 import com.hypixel.hytale.server.core.command.system.arguments.system.RequiredArg;
@@ -45,7 +44,7 @@ public class OpAddChunkAmountCommand extends AbstractAsyncCommand {
         }
 
         party.setOverride(new PartyOverride(PartyOverrides.CLAIM_CHUNK_AMOUNT, new PartyOverride.PartyOverrideValue("integer", party.getMaxClaimAmount() + selectedAmount)));
-        ClaimManager.getInstance().markDirty();
+        ClaimManager.getInstance().saveParty(party);
         sender.sendMessage(CommandMessages.MODIFIED_MAX_CHUNK_AMOUNT.param("party_name", party.getName()).param("amount", party.getMaxClaimAmount()));
         return CompletableFuture.completedFuture(null);