+++ src/main/java/com/buuz135/simpleclaims/commands/subcommand/party/op/OpModifyChunkAmountCommand.java
@@ -6,7 +6,6 @@ import com.buuz135.simpleclaims.claim.party.PartyOverrides;
 import com.buuz135.simpleclaims.commands.CommandMessages;
 import com.hypixel.hytale.component.Ref;
 import com.hypixel.hytale.component.Store;
-import com.hypixel.hytale.protocol.GameMode;
 import com.hypixel.hytale.server.core.command.system.CommandContext;
 import com.hypixel.hytale.server.core.command.system.CommandSender;
 import com.hypixel.hytale.server.core.command.system.arguments.system.RequiredArg;
@@ -56,7 +55,7 @@ public class OpModifyChunkAmountCommand extends AbstractAsyncCommand {
                         }
                         var selectedAmount = amount.get(commandContext);
                         party.setOverride(new PartyOverride(PartyOverrides.CLAIM_CHUNK_AMOUNT, new PartyOverride.PartyOverrideValue("integer", selectedAmount)));
-                        ClaimManager.getInstance().markDirty();
+                        ClaimManager.getInstance().saveParty(party);
                         player.sendMessage(CommandMessages.MODIFIED_MAX_CHUNK_AMOUNT.param("party_name", party.getName()).param("amount", selectedAmount));
                     }
                 }, world);